@@ -1,21 +1,21 @@
 {
   "name": "api",
   "version": "1.0.0",
-  "description": "Maintenance Tracker App is an application that provides users with the ability to reach out to operations or repairs department regarding repair or maintenance requests and monitor the status of the Api for an\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[C\u001b[",
+  "description": "api for a maintanance app",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon --exec babel-node index.js"
+    "start": "nodemon --exec babel-node src/index.js"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Mohamilr/Maintenance-Tracker.git"
   },
   "keywords": [
-    "tools",
+    "service",
     "request",
     "repair"
   ],
-  "author": "ibrahim mohammed",
+  "author": "Ibrahim Mohammed Damilola",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/Mohamilr/Maintenance-Tracker/issues"
@@ -26,8 +26,13 @@
     "express": "^4.17.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "eslint": "^5.16.0",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/node": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "eslint": "^6.0.1",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.18.0",
     "nodemon": "^1.19.1"
   }
 }
